+++ thread/MsgQueue.tcc
@@ -28,7 +28,7 @@ MsgQueue<_elt_t>::~MsgQueue()
     if (size() != 0)
     {
         log_err("not empty at time of destruction, size=%d",
-                queue_.size());
+                (u_int)queue_.size());
     }
     
     delete lock_;
